fixup templatize PUT
With data_source_name_query true and put_create true, use the same id-finding code
in Create as has been used in Read.

Reason: we assume API is a REST proper, and it means that PUT (used in Create instead
of POST) requires an ID. Since we are under data_source_name_query, we can assume that
the Read works - we can query by name from GETALL and obtain a good ID. Re-use the same
code. (Almost the same - Create needs to deal with IsUnknown, not IsNull).