handle state during truncateAt/createAt
Remove the subitems from the state immediately after their bulk-DELETE
succeeds. Do not remove any subitems if bulk-DELETE fails.

Reason: experiment with the most common usecase shows that bulk-DELETE
of accessrules keeps all the rules on failure. This usecase is when
one of the accessrules disappears (manually removed outside of terraform),
and thus the entire bulk fails.